refactor: move votingNeuronSelectStore to voting (#2296)
# Motivation

Move votingNeuronSelectStore to voting store file because it's directly
related to the voting.

# Tests

- no logic changes
