+++ frontend/src/tests/lib/components/proposal-detail/VotingCard/VotingCard.spec.ts
@@ -6,7 +6,7 @@ import VotingCard from "$lib/components/proposal-detail/VotingCard/VotingCard.sv
 import { SECONDS_IN_YEAR } from "$lib/constants/constants";
 import { authStore } from "$lib/stores/auth.store";
 import { neuronsStore } from "$lib/stores/neurons.store";
-import { votingNeuronSelectStore } from "$lib/stores/proposals.store";
+import { votingNeuronSelectStore } from "$lib/stores/vote-registration.store";
 import {
   SELECTED_PROPOSAL_CONTEXT_KEY,
   type SelectedProposalContext,
